feat(investment): add decimal input support and migrate to Supabase PostgreSQL
- Add initialQuantityDecimal and initialCostDecimal fields
- Add frontend FormSelect component for better form handling
- Update database configuration for Supabase PostgreSQL (port 5432)
- Add connection pool settings for Supabase optimization
- Update investment form to support decimal inputs for better UX
- Migrate from MySQL to PostgreSQL with GORM dialect changes
- Update environment example with Supabase configuration

+++ src/go-backend/cmd/migrate-investment-wallet-balance/main.go
@@ -30,7 +30,7 @@ func main() {
 
 	// Get all INVESTMENT type wallets
 	var wallets []models.Wallet
-	if err := db.DB.WithContext(ctx).Where("type = ?", v1.WalletType_INVESTMENT).Find(&wallets).Error; err != nil {
+	if err := db.DB.WithContext(ctx).Where("type = ?", int32(v1.WalletType_INVESTMENT)).Find(&wallets).Error; err != nil {
 		log.Fatalf("Failed to fetch investment wallets: %v", err)
 	}
